requirements.txt
python-dotenv==0.19.0
black==21.7b0
pylint==2.9.6
requests==2.26.0
pandas==1.3.1